requirements.txt
ultralytics>=8.0.0
pyyaml>=6.0.1
numpy>=1.21.0,<1.22.0  # Compatible with numba 0.55.1
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
opencv-python>=4.5.0
tqdm>=4.65.0
matplotlib>=3.5.0
seaborn>=0.12.0 